refactor: option item mapper 위치 이동
