+++ src/main/java/until/the/eternity/auctionhistory/domain/mapper/OpenApiItemOptionMapper.java
@@ -2,7 +2,7 @@ package until.the.eternity.auctionhistory.domain.mapper;
 
 import org.mapstruct.Mapper;
 import org.mapstruct.Mapping;
-import until.the.eternity.auctionhistory.domain.dto.external.OpenApiItemOptionResponse;
+import until.the.eternity.itemoption.domain.dto.external.OpenApiItemOptionResponse;
 import until.the.eternity.itemoption.domain.entity.ItemOption;
 
 @Mapper(componentModel = "spring")
