+++ src/main/java/until/the/eternity/itemoption/domain/dto/external/OpenApiItemOptionResponse.java
@@ -1,4 +1,4 @@
-package until.the.eternity.auctionhistory.domain.dto.external;
+package until.the.eternity.itemoption.domain.dto.external;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
